Update: Bull Market (#40)
Describes the condition of a financial market in which prices are rising or expected to rise. This term usually refers to the stock market, but can be applied to anything traded. Included are bonds, real estate, commodities, and cryptocurrencies.  Since prices continuously rise and fall during trading, the classification of "bull market" is reserved for extended amounts of time where a large portion of prices are rising.