Update: Bitcoin Improvement Proposals (BIPS)
| 👷 | Bitcoin Improvement Proposals (BIPS) | The BIP process was developed and introduced by early Bitcoin developer Amir Taaki who believed that the Bitcoin development process would benefit from becoming more structured and accountable.
Taaki submitted the first BIP (BIP 0001) on August 19, 2011, which described the BIP process. A Bitcoin Improvement Proposal (BIP) is a standard for proposing changes to the Bitcoin protocol or source for information for the Bitcoin community. It can also include changes like soft, and hard fork protocol upgrades as well as other changes that benefit from coordination across different Bitcoin software implementations, such as to the peer-to-peer layer or new backup seed formats. Not all changes to a Bitcoin software implementation affect the Bitcoin protocol some changes however make the code run more efficient or change the user interface, and these changes do not require a BIP. |